Move psi visualizer test to separate "psi" package
