+++ compiler/visualizer/tests/org/jetbrains/kotlin/visualizer/psi/PsiVisualizerForUncommonCasesGenerated.java
@@ -3,7 +3,7 @@
  * Use of this source code is governed by the Apache 2.0 license that can be found in the license/LICENSE.txt file.
  */
 
-package org.jetbrains.kotlin.visualizer;
+package org.jetbrains.kotlin.visualizer.psi;
 
 import com.intellij.testFramework.TestDataPath;
 import org.jetbrains.kotlin.test.JUnit3RunnerWithInners;
